config: add sslmode `verify-ca` and `verify-full`
When a connection is established, the added modes are treated in the
same way as the existing `require` mode as they both require a TLS
connection.

It's the responsibility of the user to configure the TLS stream to match
the semantics of Postgres client (e.g. enable peer cert verification).