tokio-postgres: use tokio mpsc for `CopyBoth` streams
The tokio implementation of mpsc channels seems to be more sophisticated
and performant than the one in futures_channel. Since the channels used
by the CopyBoth streams transfer a large amount of data during
replication this PR switches the implementation to tokio.

This PR doesn't yet alter the size of the channels in order to do an
apples to apples comparison. A subsequent test should be performed with
larger buffers between the producer and consumer tasks.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>